created a basic_block_list file and its methods, and included it in function, need to modify function, etc..
